split balance sql part into its own file - also reworked balan.ts as per recomenation
