+++ apps/builder/src/app/balance.ts
@@ -1,6 +1,6 @@
 import { firestoreUtils } from './converter/firestore-utils';
 import { firebaseApp } from './firebase';
-import { readBalance} from './mysql/account';
+import { readBalance} from './mysql/accountbalance';
 
 export const buildBalance= async (account: number): Promise<number> => {
   const db = firebaseApp.datebase;
@@ -9,7 +9,7 @@ export const buildBalance= async (account: number): Promise<number> => {
   const players = db.collection('players');
 
   return db.runTransaction(transaction => {
-      rows.forEach(docKey =>   transaction.set(players.doc(docKey.uid).update({balance: docKey.amount}), firestoreUtils.convertTimestamps(docKey)));
+      rows.forEach(({uid, amount}) => transaction.set(players.doc(uid).update({balance: amount}), firestoreUtils.convertDateTime()));
       return Promise.resolve(rows.length);
-    });
+    }); 
 }
